get_time_series now sets correctly (#3)
* get_time_series now sets correctly

get_time_series now correctly sets default values. Previous version only sets the value once the program is initiated and not everytime the function is called. So if a service ran this on two seperate occasions the default .now will be the same, even if it was 2 days apart.

* Update eloverblik.py

Spelling

Co-authored-by: Jonas Pedersen <jonas@jope.eu>